Add DependsOn Requirements to cloudformation
Some cloudformation resources depend on other resources before they
can be created or deleted. This especially causes problems when deleting
VPC's, meaning that the delete fails since there are dependent resources
still in the process of being removed form the VPC.

By adding the DependsOn attribute to these resources, this problem can be resolved.
This change uses the AWS guidance on when this is required to add this attribute to resources
that do, or may require it. The details are in the link below.
http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-attribute-dependson.html#gatewayattachment

Following the advice in the document above, the resources below are considered,

Already has DependsOn the VPCGatewayAttachment
* Amazon VPC routes that include the Internet gateway

Added DependsOn the VPCGatewayAttachment since EC2 instances have public ips
* Auto Scaling groups
* Elastic Load Balancers

Added DependsOn the VPCGatewayAttachment since they can have public ips
* RDS instances

No action required
* Elastic IP addresses - we dont have any yet
* Amazon EC2 instances - these are handled by the ASG

(Closes #177)